+++ src/socketio_transport_polling.erl
@@ -329,24 +329,21 @@ jsonp_index_digits(_) ->
     false.
 
 cors_headers(Headers, Sup) ->
-    case proplists:get_value('Origin', Headers) of
+    case socketio_listener:verify_origin_headers(Headers, socketio_listener:origins(Sup)) of
 	undefined ->
 	    {undefined, []};
-	Origin ->
-	    case socketio_listener:verify_origin(Origin, socketio_listener:origins(Sup)) of
-		true ->
-		    Headers0 = [{"Access-Control-Allow-Origin", "*"}],
-		    Headers1 =
-			case proplists:get_value('Cookie', Headers) of
-			    undefined ->
-				Headers0;
-			    _Cookie ->
-				[{"Access-Control-Allow-Credentials", "true"}|Headers0]
-			end,
-		    {true, [Headers1]};
-		false ->
-		    {false, Headers}
-	    end
+	true ->
+	    Headers0 = [{"Access-Control-Allow-Origin", "*"}],
+	    Headers1 =
+		case proplists:get_value('Cookie', Headers) of
+		    undefined ->
+			Headers0;
+		    _Cookie ->
+			[{"Access-Control-Allow-Credentials", "true"}|Headers0]
+		end,
+	    {true, [Headers1]};
+	false ->
+	    {false, Headers}
     end.
 
 reset_duration({TimerRef, Time}) ->